+++ pkg/api/vTemplate.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"errors"
 	"fmt"
 	"github.com/gin-gonic/gin"
 	"gomessage/pkg/models"
@@ -15,7 +16,11 @@ import (
 // @Router /api/v1/:namespace/template [GET]
 func ListTemplate(g *gin.Context) {
 	ns := g.Param("namespace")
-	listTemplate, _ := models.ListTemplate(ns)
+	listTemplate, err := models.ListTemplate(ns)
+	if err != nil {
+		g.JSON(http.StatusInternalServerError, utils.ResponseFailure("查询失败", err))
+		return
+	}
 	g.JSON(http.StatusOK, utils.ResponseSuccessful("数据查询成功", listTemplate))
 }
 
@@ -29,19 +34,33 @@ func PostTemplate(g *gin.Context) {
 		Namespace:    ns,
 		TemplateName: ns,
 	}
-	g.ShouldBindJSON(&body)
-	template := UpdateAddTemp(ns, body)
+	if err := g.ShouldBindJSON(&body); err != nil {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("请求内容错误", err))
+		return
+	}
+	template, err := UpdateAddTemp(ns, body)
+	if err != nil {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("模板添加失败", err))
+		return
+	}
 	g.JSON(http.StatusOK, utils.ResponseSuccessful("模板添加成功", &template))
 }
 
-func UpdateAddTemp(ns string, body models.Template) models.Template {
-	//删除指定namespace中的模板
-	listTemps, _ := models.ListTemplate(ns)
-	for _, temp := range *listTemps {
-		models.DeleteTemplate(temp.ID)
+func UpdateAddTemp(ns string, body models.Template) (models.Template, error) {
+	listTemps, err := models.ListTemplate(ns)
+	if err != nil {
+		return models.Template{}, err
 	}
-	template, _ := models.AddTemplate(&body)
-	return *template
+	for _, temp := range *listTemps {
+		if _, err := models.DeleteTemplate(temp.ID); err != nil {
+			return models.Template{}, err
+		}
+	}
+	template, err := models.AddTemplate(&body)
+	if err != nil {
+		return models.Template{}, err
+	}
+	return *template, nil
 }
 
 // GetTemplate
@@ -49,13 +68,21 @@ func UpdateAddTemp(ns string, body models.Template) models.Template {
 // @Summary 查询一个消息模板
 // @Router /api/v1/:namespace/template/:id [GET]
 func GetTemplate(g *gin.Context) {
-	id, _ := strconv.Atoi(g.Param("id"))
+	id, err := strconv.Atoi(g.Param("id"))
+	if err != nil {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("参数错误", err))
+		return
+	}
 	result, err := models.GetTemplateById(id)
 	if err != nil {
-		g.JSON(http.StatusBadRequest, "参数错误")
-	} else {
-		g.JSON(http.StatusOK, result)
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("查询失败", err))
+		return
 	}
+	if result.Namespace != g.Param("namespace") {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("参数错误", errors.New("模板不属于当前命名空间")))
+		return
+	}
+	g.JSON(http.StatusOK, utils.ResponseSuccessful("查询成功", result))
 }
 
 // PutTemplate
@@ -63,15 +90,32 @@ func GetTemplate(g *gin.Context) {
 // @Summary 修改一个消息模板
 // @Router /api/v1/:namespace/template/:id [PUT]
 func PutTemplate(g *gin.Context) {
-	id, _ := strconv.Atoi(g.Param("id"))
+	id, err := strconv.Atoi(g.Param("id"))
+	if err != nil {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("参数错误", err))
+		return
+	}
+	exist, err := models.GetTemplateById(id)
+	if err != nil {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("查询失败", err))
+		return
+	}
+	if exist.Namespace != g.Param("namespace") {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("参数错误", errors.New("模板不属于当前命名空间")))
+		return
+	}
 	body := models.Template{}
-	g.ShouldBindJSON(&body)
+	if err = g.ShouldBindJSON(&body); err != nil {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("请求内容错误", err))
+		return
+	}
+	body.Namespace = g.Param("namespace")
 	result, err := models.UpdateTemplate(id, &body)
 	if err != nil {
-		g.JSON(http.StatusBadRequest, err)
-	} else {
-		g.JSON(http.StatusOK, result)
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("更新失败", err))
+		return
 	}
+	g.JSON(http.StatusOK, utils.ResponseSuccessful("修改成功", result))
 }
 
 // DeleteTemplate
@@ -79,11 +123,24 @@ func PutTemplate(g *gin.Context) {
 // @Summary 删除一个消息模板
 // @Router /api/v1/:namespace/template/:id [DELETE]
 func DeleteTemplate(g *gin.Context) {
-	id, _ := strconv.Atoi(g.Param("id"))
+	id, err := strconv.Atoi(g.Param("id"))
+	if err != nil {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("参数错误", err))
+		return
+	}
+	exist, err := models.GetTemplateById(id)
+	if err != nil {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("查询失败", err))
+		return
+	}
+	if exist.Namespace != g.Param("namespace") {
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("参数错误", errors.New("模板不属于当前命名空间")))
+		return
+	}
 	num, err := models.DeleteTemplate(id)
 	if err != nil {
-		g.JSON(http.StatusBadRequest, err)
-	} else {
-		g.JSON(http.StatusOK, fmt.Sprintf("受影响的行数：%v", num))
+		g.JSON(http.StatusBadRequest, utils.ResponseFailure("删除失败", err))
+		return
 	}
+	g.JSON(http.StatusOK, utils.ResponseSuccessful("删除成功", fmt.Sprintf("受影响的行数：%v", num)))
 }
